bug fix
the way that the thing checked for already existing edges was ok, but if
it found a duplicate, it would not append the new edge, which is not
correct. it should delete the old edge and append the new one, thereby
"refreshing" the life of the edge back to 60 seconds.